Added alphabetical sorting script
A new PowerShell script has been added that organizes files and directories by moving them into folders named after their initial character. The script creates directories for each uppercase letter (A-Z) and moves items that start with those letters into the corresponding directory. Items that do not start with an uppercase letter are moved into a directory named "0-9". This change also includes an update to the README file to reflect this addition under the Utility category.